refactor: :construction: mapping de speaker avec workshop au lieu de Edition
